+++ vscode/src/completions/context/retrievers/tsc/ts-internals.ts
@@ -1,3 +1,4 @@
+// Copy-pasted code from scip-typescript
 import * as ts from 'typescript'
 
 // Functions in this file are based directly off corresponding functions
